# [CHANGE] regexp . to match \n and optimize performance