From: https://github.com/prometheus/prometheus/issues/7859

Improve regexp matches newlines as well


- Example with `.*` 
Before fix benchmark:

```
BenchmarkQuerier/Head/PostingsForMatchers/n="1",_j="foo"-8          1634121      789.5 ns/op
BenchmarkQuerier/Head/PostingsForMatchers/n="1",_i=~".*",_j="foo"-8       169      6567919 ns/op

```

After fix:

```
BenchmarkQuerier/Head/PostingsForMatchers/n="1",_j="foo"-8    1511605     746.8 ns/op
BenchmarkQuerier/Head/PostingsForMatchers/n="1",_i=~".*",_j="foo"-8    1561862      827.1 ns/op
```

There are already some test of .* regexp: https://github.com/prometheus/prometheus/blob/91c05eed84c296089690cd0fbb7bac5976af6e53/tsdb/querier_test.go#L2799