@jan--f 2024-07-25
No strong opinion on the feature flag, though in https://github.com/prometheus/prometheus/pull/13904 we didn't add one either.