@bboreham 2024-07-29
I absolutely think we should compile `.*` to a no-op.  That is what everybody expects.