@roidelapluie 2024-07-29
I understand but can we make sure this happens everywhere then? In relabeling as well?