@bboreham 2024-07-29
That is a good point - relabeling uses a different code path. 
https://github.com/prometheus/prometheus/blob/d699dc3c7706944aafa56682ede765398f925ef0/model/relabel/relabel.go#L173

I think changing the preamble to `^(?s:` should work.

There is a similar expression in `NewFastRegexMatcher` and in `evalLabelReplace`.