@jan--f 2024-07-30
If I'm not mistaken, the relabeling has been discussed before in https://github.com/prometheus/prometheus/pull/12180. Should this be resurrected?