@bboreham 2024-09-13
I pushed a few commits expressing what I think it should do: https://github.com/prometheus/prometheus/compare/main...bboreham:pull-14505