@bboreham 2024-09-16
Thanks for highlighting these questions.

I believe the important part is from the original issue title "Dot in regexp should also match newlines".
That should mean everywhere. ".*" is also in the title, but in parenthesis.

If you think there could be broad disagreement, we should call for comment on the Prometheus-dev mailing list.

Tests need to be fixed up where they now give the wrong answer.

I'm heading out right now but will take a look at the benchmark later.  In general we don't want special-cases in the code; the aim should be to make the code broadly readable with good performance for most cases.