@bboreham 2024-09-16
I think the mistake re the benchmark is the one special-case should be `^.*$`, since that is what Prometheus uses when the user types `.*`.