@bboreham 2024-09-16
I checked: if the user writes `foo=~".*"` then the matcher string is `.*`.
Where does the need come from to optimize `^.*$` ?