@marioferh 2024-09-16
> I checked: if the user writes `foo=~".*"` then the matcher string is `.*`. Where does the need come from to optimize `^.*$` ?
> 
> EDIT: I think it's from here: [aff9f7a#diff-19b59121b90b212f7b7da0adcd7429f2b2717fa12f4946299c25cd2b6f1b6b2eR87](https://github.com/prometheus/prometheus/commit/aff9f7a9e8f59400f4604ad40273f3cc211336b3#diff-19b59121b90b212f7b7da0adcd7429f2b2717fa12f4946299c25cd2b6f1b6b2eR87) That strikes me as a weird oddity in the benchmark - it prints out the matcher as `.*`.

It prints .*, which is correct, but the test uses ^.*$  https://github.com/prometheus/prometheus/blob/aff9f7a9e8f59400f4604ad40273f3cc211336b3/tsdb/querier_bench_test.go#L107 
Maybe this is because, in the past, all regexs were anchored? https://github.com/prometheus/docs/pull/377
I don't know why most of the test are in the anchored way.

The anchor is added https://github.com/prometheus/prometheus/blob/6febfbb3befffe6dd7efb21d3740c0261d9e6604/model/labels/regexp.go#L72:

but the comparison is done on the value, not the anchored regex:
![image](https://github.com/user-attachments/assets/7461be0c-bde2-46fd-b034-4bbb0c0063ae)

So, if we can optimize for .*, we can simply add a check for it and update the test, as was in the very beginning of the PR:
	`case m.Type == labels.MatchRegexp && m.Value == ".*":`
Otherwise, I can’t think of another approach.