@bboreham 2024-09-16
> Maybe this is because, in the past, all regexs were anchored? 

They are still anchored. When you write `=~".*"`, Prometheus turns it into `^.*$`.