@marioferh 2024-09-17
> I created #14911 to clean up the benchmarks.

Thanks, back to ".*" and waiting for the clean up

> > Maybe this is because, in the past, all regexs were anchored?
> 
> They are still anchored. When you write `=~".*"`, Prometheus turns it into `^.*$`.

And where does that happen? I see the anchor here:
https://github.com/prometheus/prometheus/blob/6febfbb3befffe6dd7efb21d3740c0261d9e6604/model/labels/regexp.go#L72

But the value here https://github.com/prometheus/prometheus/blob/d697f514dffefece99f5e9832b63c4d29383e413/tsdb/querier.go#L257 that I see is the same ".*"