@marioferh 2024-09-17
> I merged #14911.
> 
> Maybe the answer to your question is to rely on the [documentation](https://prometheus.io/docs/prometheus/latest/querying/basics/#instant-vector-selectors)?
> 
> > Regex matches are fully anchored. A match of `env=~"foo"` is treated as `env=~"^foo$"`.
> 
> "Treated as" is an internal implementation detail. When you examine the matcher it contains the original expression.

Ok thanks. PR has been rebased and title modified